entities refactor
