+++ api/admin.py
@@ -1,7 +1,10 @@
 from django.contrib import admin
 
 # Register your models here.
-from api.models import Company, Transaction
+from api.models import (
+    Company,
+    Transaction,
+)
 
 
 admin.site.register(Company)
