+++ api/management/commands/import_csv.py
@@ -1,7 +1,10 @@
 import csv
 from django.core.management.base import BaseCommand, CommandError
 
-from api.models import Company, Transaction
+from api.models import (
+    Company,
+    Transaction,
+)
 
 
 class Command(BaseCommand):
